app/requirements.txt
python-dateutil
requests
logging
sqlalchemy
python-dotenv
mysql-connector-python 
